documents: reorganize serializers
* Adds new directory.
* Moves each serializers type into new files.
